stash-migration/requirements.txt
stashy==0.7
GitPython==3.1.30